+++ ai_platform_engineering/agents/argocd/client_bindings/a2a/a2a_client_tool.py
@@ -4,7 +4,7 @@
 from langgraph.prebuilt import create_react_agent
 from cnoe_agent_utils import LLMFactory
 
-from ai_platform_engineering.agents.argocd.a2a_agent_client.agentcard import (
+from ai_platform_engineering.agents.argocd.agentcard import (
     ARGOCD_AGENT_DESCRIPTION,
     argocd_agent_card,
     argocd_agent_skill,
